slack_bot/requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
python-dotenv
dnspython